backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.10.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-multipart==0.0.12
google-generativeai==0.8.3
Pillow==11.0.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
resend==2.4.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
aiosmtplib==3.0.2
email-validator==2.2.0
